requirements.txt
redis==5.2.0
pandas==2.2.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
bcrypt==4.2.1
Flask==3.1.0
flask-socketio==5.4.1
pyjwt==2.10.1